pyproject.toml
[project]
name = "diplom"
version = "0.1.0"
description = "Identification of latent EEG state trajectories based on deep learning methods."
authors = [{name = "Ilya Rohin", email = "rohinia@my.msu.ru"}]
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "mne",
    "numpy",
    "pandas",
    "plotly",
    "scikit-learn",
    "umap-learn",
    "matplotlib",
    "ipython",
    "ipykernel",
    "nbformat>=4.2.0",
    "ipywidgets"
]

[project.optional-dependencies]
dev = [
    "ruff==0.14.0",
    "mypy==1.17.0",
    "pre-commit==4.5.0",
    "ipdb",
]

[tool.ruff]

src = ["."]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    "__pycache__",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.9
target-version = "py39"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F", "I", "B"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

# mypy global options:

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true
show_error_codes = true
warn_return_any = true
warn_unused_configs = true
exclude = [
    'dr_plotting\.ipynb',
    '^file1\.py$',  # TOML literal string (single-quotes, no escaping necessary)
    "^file2\\.py$",  # TOML basic string (double-quotes, backslash and other characters need escaping)
]

# mypy per-module options:

# [[tool.mypy.overrides]]
# module = "mycode.foo.*"
# disallow_untyped_defs = true

# [[tool.mypy.overrides]]
# module = "mycode.bar"
# warn_return_any = false

# [[tool.mypy.overrides]]
# module = [
#     "somelibrary",
#     "some_other_library"
# ]
# ignore_missing_imports = true
